[kitsu] 调整部署脚本参数

+++ build_script/deployment_server.ps1
@@ -1,4 +1,8 @@
-﻿#Set-Item WSMan:\localhost\Client\TrustedHosts -Value '*'
+﻿param (
+    [switch]$CopyServer
+)
+
+#Set-Item WSMan:\localhost\Client\TrustedHosts -Value '*'
 #Set-Item WSMan:\localhost\Client\Auth\Basic -Value True
 #Set-Item WSMan:\localhost\Service\Auth\Basic -Value True
 #Set-Item WSMan:\localhost\Client\AllowUnencrypted -Value True
@@ -39,7 +43,7 @@ Invoke-Command -ComputerName 192.168.40.181 -Credential $Credential -Authenticat
     #    Compare-Object -ReferenceObject (Get-Content -Path "D:\tmp\bin\file_association_http.exe") -DifferenceObject (Get-Content -Path "D:\kitsu\bin\file_association_http.exe")
     $Target = "D:\kitsu"
     $Tmp = "D:\tmp"
-    if ((Get-FileHash "$Target\bin\doodle_kitsu_supplement.exe").Hash -ne (Get-FileHash "$Tmp\bin\doodle_kitsu_supplement.exe").Hash)
+    if ($CopyServer -and (Get-FileHash "$Target\bin\doodle_kitsu_supplement.exe").Hash -ne (Get-FileHash "$Tmp\bin\doodle_kitsu_supplement.exe").Hash)
     {
         Stop-Service -Force -Name doodle_kitsu_supplement
         &robocopy "$Tmp\bin" "$Target\bin" /MIR
